Cast nr_quicker_queries and nr_slower_queries in simple_report plugin from tbb::atomic<uint64_t> to uint64_t before printing them out. It helped us to avoid segmentation fault during reporting execution results.

+++ percona_playback/simple_report/simple_report.cc
@@ -135,7 +135,8 @@ public:
            boost::posix_time::to_simple_string(expected_duration).c_str()
            );
     printf("%"PRIu64 " queries were quicker than expected, %"PRIu64" were slower\n",
-           nr_quicker_queries, nr_slower_queries);
+           static_cast<uint64_t>(nr_quicker_queries),
+           static_cast<uint64_t>(nr_slower_queries));
 
     printf("A total of %" PRIu64 " queries had errors.\n",
 	   uint64_t(nr_error_queries));
